Fix MSVC narrowing conversion warnings (C4244, C4267, C4305, C4838)
- icc.cpp: use float literals in Matrix3x3 aggregate initializer (C4838)
- icc.cpp: explicit cast for strlen result assigned to uint32_t (C4267)
- jpegdecoderhelper.cpp: explicit cast for std::ceil result assigned to
  unsigned int (C4244)
- jpegencoderhelper.cpp: explicit casts for std::ceil and strlen results
  assigned to unsigned int (C4244, C4267)
- CMakeLists.txt: remove /wd4244, /wd4267, /wd4305, /wd4838 suppressions

+++ lib/src/jpegencoderhelper.cpp
@@ -189,10 +189,10 @@ uhdr_error_info_t JpegEncoderHelper::encode(const uint8_t* planes[3], const unsi
     for (int i = 0; i < cinfo.num_components; i++) {
       cinfo.comp_info[i].h_samp_factor = factors[i * 2];
       cinfo.comp_info[i].v_samp_factor = factors[i * 2 + 1];
-      mPlaneWidth[i] =
-          std::ceil(((float)cinfo.image_width * cinfo.comp_info[i].h_samp_factor) / factors[6]);
-      mPlaneHeight[i] =
-          std::ceil(((float)cinfo.image_height * cinfo.comp_info[i].v_samp_factor) / factors[7]);
+      mPlaneWidth[i] = static_cast<unsigned int>(
+          std::ceil(((float)cinfo.image_width * cinfo.comp_info[i].h_samp_factor) / factors[6]));
+      mPlaneHeight[i] = static_cast<unsigned int>(
+          std::ceil(((float)cinfo.image_height * cinfo.comp_info[i].v_samp_factor) / factors[7]));
     }
     if (format != UHDR_IMG_FMT_24bppRGB888) cinfo.raw_data_in = TRUE;
     cinfo.dct_method = JDCT_ISLOW;
@@ -207,7 +207,8 @@ uhdr_error_info_t JpegEncoderHelper::encode(const uint8_t* planes[3], const unsi
       snprintf(comment, sizeof comment,
                "Source: google libuhdr v%s, Coder: libjpeg v%d, Attrib: GainMap Image",
                UHDR_LIB_VERSION_STR, JPEG_LIB_VERSION);
-      jpeg_write_marker(&cinfo, JPEG_COM, reinterpret_cast<JOCTET*>(comment), strlen(comment));
+      jpeg_write_marker(&cinfo, JPEG_COM, reinterpret_cast<JOCTET*>(comment),
+                        static_cast<unsigned int>(strlen(comment)));
     }
     if (format == UHDR_IMG_FMT_24bppRGB888) {
       while (cinfo.next_scanline < cinfo.image_height) {
